Moved all the connection-related stuff in the API to its own package, as it is not specific to wires
