+++ src/main/java/com/bluepowermod/api/connect/IConfluentConnection.java
@@ -1,4 +1,4 @@
-package com.bluepowermod.api.wire;
+package com.bluepowermod.api.connect;
 
 import java.util.Collection;
