+++ src/main/java/com/bluepowermod/api/connect/IConnectionListener.java
@@ -1,4 +1,4 @@
-package com.bluepowermod.api.wire;
+package com.bluepowermod.api.connect;
 
 public interface IConnectionListener {
