+++ src/main/java/com/bluepowermod/api/wire/redstone/IBundledDevice.java
@@ -3,9 +3,9 @@ package com.bluepowermod.api.wire.redstone;
 import net.minecraftforge.common.util.ForgeDirection;
 import uk.co.qmunity.lib.vec.IWorldLocation;
 
+import com.bluepowermod.api.connect.ConnectionType;
+import com.bluepowermod.api.connect.IConnectionCache;
 import com.bluepowermod.api.misc.MinecraftColor;
-import com.bluepowermod.api.wire.ConnectionType;
-import com.bluepowermod.api.wire.IConnectionCache;
 
 public interface IBundledDevice extends IWorldLocation {
