+++ src/main/java/com/bluepowermod/redstone/BundledConnection.java
@@ -2,8 +2,8 @@ package com.bluepowermod.redstone;
 
 import net.minecraftforge.common.util.ForgeDirection;
 
-import com.bluepowermod.api.wire.ConnectionType;
-import com.bluepowermod.api.wire.IConnection;
+import com.bluepowermod.api.connect.ConnectionType;
+import com.bluepowermod.api.connect.IConnection;
 import com.bluepowermod.api.wire.redstone.IBundledDevice;
 
 public class BundledConnection implements IConnection<IBundledDevice> {
