+++ src/main/java/com/bluepowermod/redstone/DummyRedstoneDevice.java
@@ -27,9 +27,9 @@ import uk.co.qmunity.lib.helper.MathHelper;
 import uk.co.qmunity.lib.helper.RedstoneHelper;
 import uk.co.qmunity.lib.vec.Vec3i;
 
+import com.bluepowermod.api.connect.ConnectionType;
+import com.bluepowermod.api.connect.IConnectionCache;
 import com.bluepowermod.api.misc.IFace;
-import com.bluepowermod.api.wire.ConnectionType;
-import com.bluepowermod.api.wire.IConnectionCache;
 import com.bluepowermod.api.wire.redstone.IRedstoneDevice;
 
 public class DummyRedstoneDevice implements IRedstoneDevice {
